bug fix & error ratio check

+++ DMM/include/common.hpp
@@ -1,3 +1,7 @@
-using dataType = int;
+using dataType = float;
 constexpr size_t subBlockSize = 32;
 constexpr size_t MAXTHREADSIZE = 1024;
+
+// using dataType = int;
+// constexpr size_t subBlockSize = 32;
+// constexpr size_t MAXTHREADSIZE = 1024;
